Move more metrics to `vm_interfaces`
