+++ core/lib/dal/src/tests/mod.rs
@@ -12,11 +12,10 @@ use zksync_types::{
     l2_to_l1_log::{L2ToL1Log, UserL2ToL1Log},
     protocol_upgrade::{ProtocolUpgradeTx, ProtocolUpgradeTxCommonData},
     snapshots::SnapshotRecoveryStatus,
-    tx::ExecutionMetrics,
     Address, Execute, K256PrivateKey, L1BatchNumber, L1BlockNumber, L1TxCommonData, L2BlockNumber,
     L2ChainId, PriorityOpId, ProtocolVersion, ProtocolVersionId, VmEvent, H160, H256, U256,
 };
-use zksync_vm_interface::{TransactionExecutionResult, TxExecutionStatus};
+use zksync_vm_interface::{ExecutionMetrics, TransactionExecutionResult, TxExecutionStatus};
 
 use crate::{
     blocks_dal::BlocksDal,
