+++ core/lib/vm_interface/src/lib.rs
@@ -8,9 +8,10 @@ pub use crate::{
         },
         inputs::{L1BatchEnv, L2BlockEnv, SystemEnv, TxExecutionMode, VmExecutionMode},
         outputs::{
-            BootloaderMemory, CompressedBytecodeInfo, CurrentExecutionState, ExecutionResult,
-            FinishedL1Batch, L2Block, Refunds, TransactionExecutionResult, TxExecutionStatus,
-            VmExecutionLogs, VmExecutionResultAndLogs, VmExecutionStatistics, VmMemoryMetrics,
+            BootloaderMemory, CompressedBytecodeInfo, CurrentExecutionState,
+            DeduplicatedWritesMetrics, ExecutionMetrics, ExecutionResult, FinishedL1Batch, L2Block,
+            Refunds, TransactionExecutionResult, TxExecutionStatus, VmExecutionLogs,
+            VmExecutionResultAndLogs, VmExecutionStatistics, VmMemoryMetrics,
         },
         tracer,
     },
