+++ core/node/api_server/src/web3/tests/mod.rs
@@ -17,7 +17,7 @@ use zksync_config::{
     GenesisConfig,
 };
 use zksync_dal::{transactions_dal::L2TxSubmissionResult, Connection, ConnectionPool, CoreDal};
-use zksync_multivm::interface::{TransactionExecutionResult, TxExecutionStatus};
+use zksync_multivm::interface::{ExecutionMetrics, TransactionExecutionResult, TxExecutionStatus};
 use zksync_node_genesis::{insert_genesis_batch, mock_genesis_config, GenesisParams};
 use zksync_node_test_utils::{
     create_l1_batch, create_l1_batch_metadata, create_l2_block, create_l2_transaction,
@@ -31,7 +31,7 @@ use zksync_types::{
     l2::L2Tx,
     storage::get_code_key,
     tokens::{TokenInfo, TokenMetadata},
-    tx::{ExecutionMetrics, IncludedTxLocation},
+    tx::IncludedTxLocation,
     utils::{storage_key_for_eth_balance, storage_key_for_standard_token_balance},
     AccountTreeId, Address, L1BatchNumber, Nonce, ProtocolVersionId, StorageKey, StorageLog,
     VmEvent, H256, U256, U64,
