+++ core/node/state_keeper/src/io/seal_logic/mod.rs
@@ -10,7 +10,7 @@ use anyhow::Context as _;
 use itertools::Itertools;
 use zksync_dal::{Connection, ConnectionPool, Core, CoreDal};
 use zksync_multivm::{
-    interface::TransactionExecutionResult,
+    interface::{DeduplicatedWritesMetrics, TransactionExecutionResult},
     utils::{
         get_max_batch_gas_limit, get_max_gas_per_pubdata_byte, ModifiedSlot,
         StorageWritesDeduplicator,
@@ -22,7 +22,7 @@ use zksync_types::{
     event::extract_long_l2_to_l1_messages,
     helpers::unix_timestamp_ms,
     l2_to_l1_log::UserL2ToL1Log,
-    tx::{tx_execution_info::DeduplicatedWritesMetrics, IncludedTxLocation},
+    tx::IncludedTxLocation,
     utils::display_timestamp,
     Address, ExecuteTransactionCommon, ProtocolVersionId, StorageKey, StorageLog, Transaction,
     VmEvent, H256,
