+++ core/node/state_keeper/src/seal_criteria/criteria/geometry_seal_criteria.rs
@@ -69,7 +69,8 @@ impl SealCriterion for CircuitsCriterion {
 }
 #[cfg(test)]
 mod tests {
-    use zksync_types::{circuit::CircuitStatistic, tx::ExecutionMetrics};
+    use zksync_multivm::interface::ExecutionMetrics;
+    use zksync_types::circuit::CircuitStatistic;
 
     use super::*;
