+++ core/node/state_keeper/src/seal_criteria/criteria/pubdata_bytes.rs
@@ -66,7 +66,7 @@ impl SealCriterion for PubDataBytesCriterion {
 
 #[cfg(test)]
 mod tests {
-    use zksync_types::tx::ExecutionMetrics;
+    use zksync_multivm::interface::ExecutionMetrics;
 
     use super::*;
