+++ core/node/state_keeper/src/types.rs
@@ -5,10 +5,8 @@ use std::{
 
 use zksync_dal::{Connection, Core, CoreDal};
 use zksync_mempool::{L2TxFilter, MempoolInfo, MempoolStore};
-use zksync_multivm::interface::VmExecutionResultAndLogs;
-use zksync_types::{
-    block::BlockGasCount, tx::ExecutionMetrics, Address, Nonce, PriorityOpId, Transaction,
-};
+use zksync_multivm::interface::{ExecutionMetrics, VmExecutionResultAndLogs};
+use zksync_types::{block::BlockGasCount, Address, Nonce, PriorityOpId, Transaction};
 
 use super::{
     metrics::StateKeeperGauges,
