+++ core/node/state_keeper/src/updates/l1_batch_updates.rs
@@ -1,7 +1,7 @@
-use zksync_multivm::interface::{FinishedL1Batch, TransactionExecutionResult};
+use zksync_multivm::interface::{ExecutionMetrics, FinishedL1Batch, TransactionExecutionResult};
 use zksync_types::{
     block::BlockGasCount, priority_op_onchain_data::PriorityOpOnchainData,
-    tx::tx_execution_info::ExecutionMetrics, ExecuteTransactionCommon, L1BatchNumber,
+    ExecuteTransactionCommon, L1BatchNumber,
 };
 
 use crate::{updates::l2_block_updates::L2BlockUpdates, utils::new_block_gas_count};
